@@ -14,6 +14,7 @@ import stark.dataworks.basic.data.redis.RedisQuickOperation;
 import stark.dataworks.basic.params.OutValue;
 import stark.dataworks.boot.autoconfig.minio.EasyMinio;
 import stark.dataworks.boot.autoconfig.web.LogArgumentsAndResponse;
+import stark.dataworks.boot.web.PaginatedData;
 import stark.dataworks.boot.web.ServiceResponse;
 import stark.stellasearch.dao.UserVideoInfoMapper;
 import stark.stellasearch.dao.UserVideoLikeMapper;
@@ -25,7 +26,7 @@ import stark.stellasearch.dto.params.*;
 import stark.stellasearch.dto.results.VideoPlayInfo;
 import stark.stellasearch.service.dto.User;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.io.IOException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
@@ -36,7 +37,8 @@ import java.util.concurrent.TimeUnit;
 @Service
 @LogArgumentsAndResponse
 @Validated
-public class VideoService {
+public class VideoService
+{
     public static final String VIDEO_UPLOAD_TASK_ID_PREFIX = "videoUpload";
     public static final String VIDEO_CHUNK_COUNT_PREFIX = "chunkCount";
     public static final String VIDEO_CHUNK_SET_PLACEHOLDER = "-1";
@@ -62,7 +64,7 @@ public class VideoService {
 
     @Autowired
     private UserVideoLikeMapper userVideoLikeMapper;
-    
+
     @Autowired
     private VideoPlayRecordMapper videoPlayRecordMapper;
 
@@ -78,11 +80,13 @@ public class VideoService {
         return taskId;
     }
 
-    private String generateTaskChunkCountKey(String taskId) {
+    private String generateTaskChunkCountKey(String taskId)
+    {
         return taskId.replace(VIDEO_UPLOAD_TASK_ID_PREFIX, VIDEO_CHUNK_COUNT_PREFIX);
     }
 
-    public ServiceResponse<String> generateNewVideoUploadingTask(@Valid NewVideoUploadingTaskRequest request) {
+    public ServiceResponse<String> generateNewVideoUploadingTask(@Valid NewVideoUploadingTaskRequest request)
+    {
         String fileExtension = request.getVideoFileExtension();
         if (!VIDEO_FILE_EXTENSION_SET.contains(fileExtension))
             return ServiceResponse.buildErrorResponse(-3, "Unacceptable video file extension: " + fileExtension);
@@ -102,7 +106,8 @@ public class VideoService {
         return ServiceResponse.buildSuccessResponse(taskId);
     }
 
-    public ServiceResponse<Boolean> uploadVideoChunk(@Valid VideoChunkUploadingRequest request) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
+    public ServiceResponse<Boolean> uploadVideoChunk(@Valid VideoChunkUploadingRequest request) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException
+    {
         // Validations:
         // 1. The task ID must exist.
         // 2. The chunk index must between 0 and (chunkCount - 1).
@@ -159,7 +164,8 @@ public class VideoService {
      * @throws XmlParserException
      * @throws InternalException
      */
-    public ServiceResponse<Long> composeVideoChunks(@Valid ComposeVideoChunksRequest request) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
+    public ServiceResponse<Long> composeVideoChunks(@Valid ComposeVideoChunksRequest request) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException
+    {
         String taskId = request.getVideoUploadingTaskId();
         if (!redisQuickOperation.containsKey(taskId))
             return ServiceResponse.buildErrorResponse(-2, "Invalid video uploading task ID: " + taskId);
@@ -190,7 +196,8 @@ public class VideoService {
 
         if (chunkCount == 1)
             easyMinio.copyObject(bucketNameVideos, sortedChunkNames.get(0), videoName);
-        else {
+        else
+        {
             int chunkNamePrefixLength = chunkNamePrefix.length();
 
             sortedChunkNames.sort((x, y) ->
@@ -218,7 +225,8 @@ public class VideoService {
      *
      * @param videoName
      */
-    private long saveUserVideoInfo(String videoName) {
+    private long saveUserVideoInfo(String videoName)
+    {
         // TODO: Validate if the video name exists in the database (i.e. in MinIO).
         // We use UUID here, so the probability of collision is relatively small.
 
@@ -238,7 +246,8 @@ public class VideoService {
         return userVideoInfo.getId();
     }
 
-    private void resetExpirationOfUploadingTask(String taskId) {
+    private void resetExpirationOfUploadingTask(String taskId)
+    {
         redisQuickOperation.expire(taskId, 30, TimeUnit.MINUTES);
     }
 
@@ -248,7 +257,8 @@ public class VideoService {
      *
      * @param taskId
      */
-    private void clearUploadingTask(String taskId, List<String> sortedChunkNames) {
+    private void clearUploadingTask(String taskId, List<String> sortedChunkNames)
+    {
         // TODO: Clear the database record.
 
         lowPriorityTaskExecutor.execute(() ->
@@ -261,7 +271,8 @@ public class VideoService {
             redisQuickOperation.delete(videoFileExtensionKey);
 
             // Remove chunks in MinIO (compose operation will only create a new object without any changes on existing objects.)
-            try {
+            try
+            {
                 easyMinio.deleteObjects(bucketNameVideos, sortedChunkNames);
             } catch (ServerException | InsufficientDataException | ErrorResponseException | IOException |
                      NoSuchAlgorithmException | InvalidKeyException | InvalidResponseException | XmlParserException |
@@ -272,7 +283,8 @@ public class VideoService {
         });
     }
 
-    public ServiceResponse<Boolean> setVideoInfo(@Valid VideoInfoFormData request) {
+    public ServiceResponse<Boolean> setVideoInfo(@Valid VideoInfoFormData request)
+    {
         OutValue<UserVideoInfo> userVideoInfoOutValue = new OutValue<>();
 
         String errorMessage = validateVideoLabels(request.getLabels());
@@ -286,12 +298,14 @@ public class VideoService {
         return updateVideoInfoResponse(request, userVideoInfoOutValue.getValue());
     }
 
-    private ServiceResponse<Boolean> updateVideoInfoResponse(VideoInfoFormData request, UserVideoInfo userVideoInfo) {
+    private ServiceResponse<Boolean> updateVideoInfoResponse(VideoInfoFormData request, UserVideoInfo userVideoInfo)
+    {
         updateVideoInfo(request, userVideoInfo);
         return ServiceResponse.buildSuccessResponse(true);
     }
 
-    private void updateVideoInfo(VideoInfoFormData request, UserVideoInfo userVideoInfo) {
+    private void updateVideoInfo(VideoInfoFormData request, UserVideoInfo userVideoInfo)
+    {
         List<Long> labels = request.getLabels();
         labels.sort(Long::compareTo);
         String labelArrayText = JsonSerializer.serialize(labels);
@@ -306,7 +320,8 @@ public class VideoService {
         userVideoInfoMapper.updateVideoInfoById(userVideoInfo);
     }
 
-    private String validateVideoInfoForInitialization(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue) {
+    private String validateVideoInfoForInitialization(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue)
+    {
         // Validations:
         // 1. The video with the specified ID exists.
         // 2. User uploaded this video.
@@ -328,11 +343,13 @@ public class VideoService {
         return null;
     }
 
-    private String validateVideoLabels(List<Long> labels) {
+    private String validateVideoLabels(List<Long> labels)
+    {
         if (CollectionUtils.isEmpty(labels))
             return "There are no video labels.";
 
-        for (Long label : labels) {
+        for (Long label : labels)
+        {
             if (label == null || label <= 0)
                 return "Invalid label ID: " + label + ". We only accept positive label IDs.";
         }
@@ -340,7 +357,8 @@ public class VideoService {
         return null;
     }
 
-    private String validateVideoInfoForCreator(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue) {
+    private String validateVideoInfoForCreator(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue)
+    {
         // Validations:
         // 1. Video exists.
         // 2. Video creator ID = current user ID.
@@ -361,7 +379,8 @@ public class VideoService {
         return null;
     }
 
-    private String validateVideoInfoForUpdate(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue) {
+    private String validateVideoInfoForUpdate(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue)
+    {
         // Validations:
         // 1. Video exists.
         // 2. Video creator ID = current user ID.
@@ -379,7 +398,8 @@ public class VideoService {
         return null;
     }
 
-    public ServiceResponse<List<VideoPlayInfo>> getVideoInfoOfCurrentUser(@Valid PaginationRequestParam request) {
+    public ServiceResponse<List<VideoPlayInfo>> getVideoInfoOfCurrentUser(@Valid PaginationRequestParam request)
+    {
         GetVideoInfosByUserIdQueryParam queryParam = new GetVideoInfosByUserIdQueryParam();
         queryParam.setUserId(UserContextService.getCurrentUser().getId());
         queryParam.setPaginationParam(request);
@@ -391,12 +411,14 @@ public class VideoService {
         return response;
     }
 
-    public ServiceResponse<Long> countVideoByUserId() {
+    public ServiceResponse<Long> countVideoByUserId()
+    {
         Long count = userVideoInfoMapper.countVideoByUserId(UserContextService.getCurrentUser().getId());
         return ServiceResponse.buildSuccessResponse(count);
     }
 
-    public ServiceResponse<Boolean> updateVideoInfo(@Valid VideoInfoFormData request) {
+    public ServiceResponse<Boolean> updateVideoInfo(@Valid VideoInfoFormData request)
+    {
         String errorMessage = validateVideoLabels(request.getLabels());
         if (errorMessage != null)
             return ServiceResponse.buildErrorResponse(-7, errorMessage);
@@ -409,7 +431,8 @@ public class VideoService {
         return updateVideoInfoResponse(request, userVideoInfoOutValue.getValue());
     }
 
-    public ServiceResponse<VideoInfoFormData> getVideoInfoFormDataById(long videoId) {
+    public ServiceResponse<VideoInfoFormData> getVideoInfoFormDataById(long videoId)
+    {
         OutValue<UserVideoInfo> userVideoInfoOutValue = new OutValue<>();
         String errorMessage = validateVideoIdExistence(videoId, userVideoInfoOutValue);
         if (errorMessage != null)
@@ -419,7 +442,8 @@ public class VideoService {
         return ServiceResponse.buildSuccessResponse(request);
     }
 
-    private static VideoInfoFormData convertToVideoInfoFormData(UserVideoInfo userVideoInfo) {
+    private static VideoInfoFormData convertToVideoInfoFormData(UserVideoInfo userVideoInfo)
+    {
         String labelIds = userVideoInfo.getLabelIds();
         List<Long> labels = JsonSerializer.deserializeList(labelIds, Long.class);
         labels.sort(Long::compareTo);
@@ -437,7 +461,8 @@ public class VideoService {
         return formData;
     }
 
-    private String validateVideoIdExistence(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue) {
+    private String validateVideoIdExistence(long videoId, OutValue<UserVideoInfo> userVideoInfoOutValue)
+    {
         if (videoId <= 0)
             return "Invalid video ID: " + videoId;
 
@@ -449,7 +474,8 @@ public class VideoService {
         return null;
     }
 
-    public ServiceResponse<VideoPlayInfo> getVideoInfoById(long videoId) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
+    public ServiceResponse<VideoPlayInfo> getVideoInfoById(long videoId) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException
+    {
         VideoPlayInfo videoPlayInfo = userVideoInfoMapper.getVideoPlayInfoById(videoId);
         if (videoPlayInfo == null)
             return ServiceResponse.buildErrorResponse(-7, "Invalid video ID: " + videoId);
@@ -474,9 +500,7 @@ public class VideoService {
         // Validate if the video is liked before
         UserVideoLike userVideoLike = userVideoLikeMapper.getUserVideoLike(UserContextService.getCurrentUser().getId(), request.getVideoId());
         if (userVideoLike != null)
-        {
             return ServiceResponse.buildSuccessResponse(true);
-        }
 
         String errorMessage = insertVideoLike(request, videoPlayInfo);
         if (errorMessage != null)
@@ -513,8 +537,10 @@ public class VideoService {
         userVideoLikeMapper.deleteLike(UserContextService.getCurrentUser().getId(), request.getVideoId());
         return ServiceResponse.buildSuccessResponse(true);
     }
+
     /**
      * Insert record to table of video play count
+     *
      * @param videoPlayInfo
      * @return
      */
@@ -532,5 +558,33 @@ public class VideoService {
             return "Insert record to table of video play count failed";
 
         return null;
+    }
+
+    public ServiceResponse<PaginatedData<VideoPlayInfo>> searchVideo(@Valid SearchVideoRequest request) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException
+    {
+        String keyword = request.getKeyword();
+        long videoCountByKeyword = userVideoInfoMapper.countVideoByKeyword(keyword);
+        if (videoCountByKeyword == 0)
+            return ServiceResponse.buildSuccessResponse(new PaginatedData<>());
+
+        GetVideoInfosByKeywordQueryParam queryParam = new GetVideoInfosByKeywordQueryParam();
+        queryParam.setKeyword(keyword);
+        queryParam.setPaginationParam(request);
+
+        List<VideoPlayInfo> videoPlayInfos = userVideoInfoMapper.getVideoPlayInfosByKeyword(queryParam);
+        for (VideoPlayInfo videoPlayInfo : videoPlayInfos)
+        {
+            String videoPlayUrl = easyMinio.getObjectUrl(bucketNameVideos, videoPlayInfo.getNameInOss());
+            videoPlayInfo.setVideoPlayUrl(videoPlayUrl);
+        }
+
+        PaginatedData<VideoPlayInfo> paginatedData = new PaginatedData<>();
+        paginatedData.setData(videoPlayInfos);
+        paginatedData.setTotal(videoCountByKeyword);
+
+        ServiceResponse<PaginatedData<VideoPlayInfo>> response = ServiceResponse.buildSuccessResponse(paginatedData);
+        response.putExtra("size", videoPlayInfos.size());
+
+        return response;
     }
 }
