@@ -9,9 +9,9 @@ import stark.dataworks.basic.data.json.JsonSerializer;
 import stark.dataworks.boot.web.ServiceResponse;
 import stark.stellasearch.service.constants.SecurityConstants;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
@@ -21,13 +21,7 @@ public class LogoutSuccessJsonHandler implements LogoutSuccessHandler
     @Override
     public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException
     {
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-
         ServiceResponse<Boolean> logoutResponse = ServiceResponse.buildSuccessResponse(true, SecurityConstants.LOGOUT_SUCCESS);
-
-        String s = JsonSerializer.serialize(logoutResponse);
-        log.info("Logout success message = {}", s);
-        response.getWriter().println(s);
-        response.flushBuffer();
+        logoutResponse.writeToResponse(response);
     }
 }
