@@ -16,13 +16,11 @@ import stark.stellasearch.service.dto.User;
 import stark.stellasearch.service.redis.RedisKeyManager;
 import stark.stellasearch.service.redis.StellaRedisOperation;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 
 @Slf4j
 public class TokenLoginFilter extends OncePerRequestFilter
@@ -32,31 +30,20 @@ public class TokenLoginFilter extends OncePerRequestFilter
     private final UserDetailsService userDetailsService;
     private final StellaRedisOperation stellaRedisOperation;
     private final RedisKeyManager redisKeyManager;
-    private final List<String> ignoreUris;
 
-    public TokenLoginFilter(JwtService jwtService, RedisQuickOperation redisQuickOperation, UserDetailsService userDetailsService, StellaRedisOperation stellaRedisOperation, String contextPath, RedisKeyManager redisKeyManager)
+    public TokenLoginFilter(JwtService jwtService, RedisQuickOperation redisQuickOperation, UserDetailsService userDetailsService, StellaRedisOperation stellaRedisOperation, RedisKeyManager redisKeyManager)
     {
         this.jwtService = jwtService;
         this.redisQuickOperation = redisQuickOperation;
         this.userDetailsService = userDetailsService;
         this.stellaRedisOperation = stellaRedisOperation;
         this.redisKeyManager = redisKeyManager;
-
-        ignoreUris = new ArrayList<>();
-        for (String uri : SecurityConstants.NON_AUTHENTICATE_URIS)
-            ignoreUris.add(contextPath + uri);
-        ignoreUris.add(contextPath + SecurityConstants.DEFAULT_LOGIN_URI);
     }
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException
     {
-        String requestURI = request.getRequestURI();
-        if (ignoreUris.contains(requestURI))
-        {
-            filterChain.doFilter(request, response);
-            return;
-        }
+        // Parse the login state if there is a token, no matter if it is a login state required Uri.
 
         String token = TokenHandler.getToken(request, SecurityConstants.SSO_COOKIE_NAME);
 
