@@ -30,6 +30,7 @@ public class SecurityConstants
                     "/account/register",
                     "/account/validate-token",
                     "/video/cover/*",
-                    "/video/play"
+                    "/video/play",
+                    "/comment/list"
             };
 }
